pyproject.toml
[project]
name = "mcp-sv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.1",
    "pymongo>=4.4.0",
    "uvicorn>=0.24.0",
    "mysql-connector-python>=8.0.0",
    "python-dotenv>=1.0.0"
]

[dependency-groups]
dev = [
    "mcp[cli]",
]


